feat: 좋아요 및 쿠폰 동시성 테스트 반영

+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductMapper.java
@@ -9,7 +9,7 @@ public class ProductMapper {
         return new Product(criteria.name(), new Stock(criteria.stockValue()), new Money(criteria.priceValue()), criteria.brandId());
     }
 
-    public static ProductResult fromProduct(Product product, String brandName, int likeCount) {
+    public static ProductResult fromProduct(Product product, String brandName, Long likeCount) {
         return new ProductResult(product.getId(),
                 product.getName(),
                 product.getStock().getValue(),
