+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductResult.java
@@ -2,7 +2,7 @@ package com.loopers.application.product;
 
 import java.math.BigDecimal;
 
-public record ProductResult(Long productId, String name, int stock, BigDecimal price, String brandName, int likeCount) {
+public record ProductResult(Long productId, String name, int stock, BigDecimal price, String brandName, Long likeCount) {
 
 }
