+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/product/ProductV1Dto.java
@@ -12,7 +12,7 @@ public class ProductV1Dto {
             Integer stock,
             BigDecimal price,
             String brandName,
-            Integer likeCount
+            Long likeCount
     ) {
         public static ProductResponse from(ProductResult info) {
             return new ProductResponse(
